requirements.txt
pytorch==1.6.0 
torchvision==0.7.0 
cudatoolkit=10.1
numpy==1.21.2
